move duration_name
